[Update]헤더 반응형 CSS추가 및 useEffect 분리
