+++ src/components/Header/Header.css
@@ -12,6 +12,16 @@ header {
     background-color: #ffffff;
 }
 
+.logoText {
+    display: none;
+}
+
+.title {
+    font-weight: 700;
+    font-size: 25px;
+    color: var(--blue);
+}
+
 .leftMenu {
     display: flex;
     align-items: center;
@@ -22,8 +32,38 @@ header {
     font-size: 18px;
     font-weight: 700;
     color: var(--gray600);
+    white-space: nowrap;
 }
 
 .leftMenu .selected {
     color: var(--blue);
+}
+
+@media (max-width: 1200px) {
+    header {
+        padding: 0 24px;
+    }
+}
+
+@media (max-width: 767px) {
+    header {
+        padding: 0 16px;
+    }
+
+    .leftMenu {
+        gap: 8px;
+    }
+
+    .logoIcon {
+        display: none;
+    }
+
+    .logoText {
+        display: block;
+    }
+
+    .leftMenu a {
+        font-size: 16px;
+        padding: 0;
+    }
 }
