+++ src/components/Header/Header.jsx
@@ -1,5 +1,8 @@
 import React from "react";
 import { Link, NavLink } from "react-router-dom";
+import logoIcon from "../../assets/icon/logoIcon.svg";
+import logoText from "../../assets/icon/logoText.svg";
+import profileIcon from "../../assets/icon/profile.svg";
 import "./Header.css";
 
 const setLinkStyle = ({ isActive }) => {
@@ -10,11 +13,9 @@ const Header = () => {
     return (
         <header>
             <div className="leftMenu">
-                <Link to="/">
-                    <img
-                        src={require("../../assets/icon/logo/logo_S.png")}
-                        alt="로고"
-                    />
+                <Link className="logo" to="/">
+                    <img className="logoIcon" src={logoIcon} alt="로고" />
+                    <img className="logoText" src={logoText} alt="로고" />
                 </Link>
                 <NavLink to="/" className={setLinkStyle}>
                     자유게시판
@@ -24,10 +25,7 @@ const Header = () => {
                 </NavLink>
             </div>
             <NavLink to="/">
-                <img
-                    src={require("../../assets/icon/profile/profile_L.png")}
-                    alt="프로필 이미지"
-                />
+                <img src={profileIcon} alt="프로필 이미지" />
             </NavLink>
         </header>
     );
