--- src/pages/ItemPage/PageButton.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const PageButton = () => {
-    return <div></div>;
-};
-
-export default PageButton;
